Converts a FLIR image into an RGB SpatRaster

Extracts the full RGB image from a FLIR radiometric JPEG, optionally crops it based on FLIR metadata to match the thermal image's dimensions, and converts it to a `SpatRaster` object. This function can save the extracted and cropped RGB image as a TIFF file. Cropping is useful to align the RGB image with the corresponding thermal image, enabling direct comparison or fusion of data from both sources.

Usage

flir_rgb_to_rast(
  img_path,
  exiftool_path,
  out_path = NULL,
  crop = TRUE,
  out_path_crop = NULL,
  overwrite = TRUE
)

Arguments

img_path

Path to the input FLIR radiometric JPEG file.

exiftool_path

Path to the EXIFtool executable, including trailing slash.

out_path

Optional. The location/path where to save the full RGB image. IF `NULL` the directory of the input will be used with a file name suffix _full_RGB.jpg. This image will be created by EXIFtool by extracting the RGB image from metadata. If a file path is specified it must be of jpg extension.

crop

Logical flag indicating whether to crop the extracted RGB image to match the dimensions of the thermal image. Defaults to TRUE.

out_path_crop

Path to where the cropped imaged will be saved. This will use `terra::writeRaster` to write the image data.

overwrite

Overwrite image data in `terra::writeRaster`? Default: TRUE

Value

A `SpatRaster` object representing the extracted image. If `crop=TRUE` the cropped RGB image will be returned instead of the full RGB image.